refactor: chatbox component naming
