+++ client/src/components/Room/Chatbox/index.ts
@@ -1,0 +1,2 @@
+import Chatbox from './Chatbox';
+export default Chatbox;
