--- client/src/components/Room/CodeExec/index.ts
@@ -1,2 +1,0 @@
-import CodeExec from './CodeExec';
-export default CodeExec;
