+++ client/src/components/Room/RoomCodeExec/index.ts
@@ -1,0 +1,2 @@
+import RoomCodeExec from './RoomCodeExec';
+export default RoomCodeExec;
