--- client/src/components/Room/RoomDrawer/index.ts
@@ -1,2 +1,0 @@
-import RoomDrawer from './RoomDrawer';
-export default RoomDrawer;
