--- client/src/components/Room/index.ts
@@ -1,2 +1,0 @@
-import Room from './Room';
-export default Room;
